feat: adopt shopspring/decimal for monetary precision (Phase C)
Replace float64 with decimal.Decimal across the entire money path:
- sqlc.yaml override: numeric → decimal.Decimal (Postgres numeric scans
  directly into decimal with full precision, no float64 rounding)
- SQL queries: cast aggregates to ::numeric (not ::float8) so sqlc infers
  decimal.Decimal for SUM/CASE/COALESCE results
- Domain types: Ledger.Amount, Employee.WageAmount, DashboardSnapshot
  monetary fields, EmployeeBalance.Balance, DailySummary.TotalWageBill,
  WageBillTrend.TotalWages, LedgerSummaryRange, TenantConfig OT fields,
  AdvanceRequest.Amount all use decimal.Decimal
- Service math: use .Add()/.Sub()/.Mul()/.Div()/.Equal() instead of
  operators; .InexactFloat64() at API boundaries for JSON compat
- Tests: updated all monetary literals and assertions to decimal types
- API contract preserved: decimal.MarshalJSON outputs JSON numbers

+++ server/database/queries/dashboard.sql
@@ -5,17 +5,17 @@ SELECT
   (SELECT COUNT(*) FROM attendance WHERE attendance.tenant_id = @tenant_id AND attendance.date = @today AND attendance.status = 'present')::int AS present,
   (SELECT COUNT(*) FROM attendance WHERE attendance.tenant_id = @tenant_id AND attendance.date = @today AND attendance.status = 'absent')::int AS absent,
   (SELECT COUNT(*) FROM attendance WHERE attendance.tenant_id = @tenant_id AND attendance.date = @today AND attendance.status IN ('paid_leave','week_off'))::int AS on_leave,
-  (SELECT COALESCE(SUM(ledger.amount),0)::float8 FROM ledger WHERE ledger.tenant_id = @tenant_id AND ledger.date = @today AND ledger.type = 'jama') AS daily_jama_total,
-  (SELECT COALESCE(SUM(CASE WHEN ledger.type='jama' THEN ledger.amount ELSE 0 END),0)::float8
+  (SELECT COALESCE(SUM(ledger.amount),0)::numeric FROM ledger WHERE ledger.tenant_id = @tenant_id AND ledger.date = @today AND ledger.type = 'jama') AS daily_jama_total,
+  (SELECT COALESCE(SUM(CASE WHEN ledger.type='jama' THEN ledger.amount ELSE 0 END),0)::numeric
      FROM ledger WHERE ledger.tenant_id = @tenant_id AND ledger.date BETWEEN @month_start AND @today) AS wage_bill_mtd,
   (SELECT (COALESCE(SUM(CASE WHEN ledger.type='udhaar' THEN ledger.amount ELSE 0 END),0)
-        - COALESCE(SUM(CASE WHEN ledger.type='jama' THEN ledger.amount ELSE 0 END),0))::float8
+        - COALESCE(SUM(CASE WHEN ledger.type='jama' THEN ledger.amount ELSE 0 END),0))::numeric
      FROM ledger WHERE ledger.tenant_id = @tenant_id) AS total_outstanding;
 
 -- name: ListEmployeeBalances :many
 SELECT ledger.employee_id,
   (COALESCE(SUM(CASE WHEN ledger.type = 'jama' THEN ledger.amount ELSE 0 END),0)
-    - COALESCE(SUM(CASE WHEN ledger.type = 'udhaar' THEN ledger.amount ELSE 0 END),0))::float8 AS balance
+    - COALESCE(SUM(CASE WHEN ledger.type = 'udhaar' THEN ledger.amount ELSE 0 END),0))::numeric AS balance
 FROM ledger
 WHERE ledger.tenant_id = @tenant_id
 GROUP BY ledger.employee_id;
